Verification and webhook modification
A bit late on this commit.

The verification term "highest" now works with the w.cverify command. This is what is shown on discord, but the library does not support it by default at the moment

Implements hookurls command and add hook command, for getting the urls for webhooks and adding webhooks to channels respectively